Making translate work
