+++ bionumpy/sequence/__init__.py
@@ -6,6 +6,7 @@ from .string_matcher import construct_wildcard_matcher, \
     construct_fixed_len_regex_matchers
 from .position_weight_matrix import get_motif_scores, _pwm_from_counts
 from .dna import get_reverse_complement, get_strand_specific_sequences
+from .translate import translate_dna_to_protein
 from .count_encoded import count_encoded
