+++ bionumpy/util/testing.py
@@ -7,12 +7,12 @@ from npstructures import RaggedArray
 import numpy as np
 
 def assert_encoded_array_equal(array1, array2):
-    # array1, array2 = (as_encoded_array(a) for a in (array1, array2))
+    array1, array2 = (as_encoded_array(a) for a in (array1, array2))
     assert array1.encoding == array2.encoding, (array1.encoding, array2.encoding)
     assert_array_equal(array1.raw(), array2.raw())
 
 def assert_encoded_raggedarray_equal(array1, array2):
-    # array1, array2 = (as_encoded_array(a) for a in (array1, array2))
+    array1, array2 = (as_encoded_array(a) for a in (array1, array2))
     assert array1.encoding == array2.encoding, (array1.encoding, array2.encoding)
     assert_raggedarray_equal(array1.raw(), array2.raw())
