+++ tests/test_translate.py
@@ -1,9 +1,8 @@
 from bionumpy.sequence.translate import Translate
-from bionumpy import as_encoded_array
+from bionumpy.util.testing import assert_encoded_array_equal, assert_encoded_raggedarray_equal
 
 
 def test_translate():
     seqs = ["TTTTTC", "GGG"]
     prots = Translate().windowed(seqs)
-    print(prots)
-    assert prots == as_encoded_array(["FF", "G"])
+    assert_encoded_raggedarray_equal(prots, ["FF", "G"])
